*BIG UPDATE*
This update converts everything from XNA 2.0 VS 2005, to XNA 3.0 VS 2008. So make sure you have all the necessary software, or else this updated version will not work. I have taken (basically) everything out of the Game1 class and seperated them into their own individual classes. In fact, we now have 7 classes that are now in use throughout the game:

Game1.cs: This class is what it always has been, but instead of everything being inside of it and having one gigantic class, everything has been broken down.

GameInput.cs: This class helps with player input. Inside of having to write if gamepad is pressed or if keyboard is key down, now you can just call this class, and say which button/direction and it'll do everything for you. For example, Up button is now W, Up Arrow, and DPAD up.

Graphics.cs: Gets the GraphiceDeviceManager. Will probably update this later on to have better naming convention.

Log.cs: This class is an I/O class used for error handling. Currently it will create a .txt file in c:\DuneBuggyLogs with the error message instead of crashing the game. Currently it is only used in the sound class, but will eventually need to be put into all of the classes.

Menu.cs: This class controls the menu aspects of the game, from updating, to drawing, and loading the content.

Play.cs: This is an example class. Eventually it will be broken down into specific Levels, but it is an example on how to load, update, and draw 2d and 3d models on a screen.

Sound.cs: This class is used to play, load, start and stop sounds.

When all is said and done, the only thing the game does is show a menu, go to a hud when it's in the play function, and if you see the HUD you can press Z and see the 3D model and rotate it by pressing the right arrow key. The sooner all the textures and 3D models are finished the faster we can start implementing them. Remember, the first level should be the hardest, everything after that can almost be repeated from that level.

Hopefully this update works, if not I still have a copy on my hard drive.